@@ -11,20 +11,22 @@
 		"php": ">=8.2.0",
 		"guzzlehttp/guzzle": "^7.0.1",
 		"intervention/image": "^2.7",
-		"laravel/framework": "^11.0",
-		"laravel/tinker": "^2.0",
+		"laravel/framework": "^11.9",
+		"laravel/octane": "^2.5",
+		"laravel/tinker": "^2.9",
+		"laravel/ui": "^4.0",
 		"maxmind-db/reader": "1.4.1",
 		"namshi/jose": "7.0.*",
+		"spatie/laravel-activitylog": "^4.10",
 		"symfony/lock": "^6.0",
-		"tymon/jwt-auth": "^2.0",
-		"laravel/ui": "^4.0"
+		"tymon/jwt-auth": "^2.2.1"
 	},
 	"require-dev": {
 		"barryvdh/laravel-debugbar": "^3.12",
 		"fakerphp/faker": "^1.23",
-		"mockery/mockery": "^1.3.1",
+		"mockery/mockery": "^1.6",
 		"nunomaduro/collision": "^8.1",
-		"phpunit/phpunit": "^10.0",
+		"phpunit/phpunit": "^11.0.1",
 		"spatie/laravel-ignition": "^2.0"
 	},
 	"autoload": {
