@@ -4,6 +4,7 @@ namespace App\Extensions\Qm\Matchup;
 
 use App\Models\Game;
 use App\Models\QmQueueEntry;
+use App\Models\QmLadderRules;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Log;
 
@@ -25,8 +26,11 @@ class TeamMatchupHandler extends BaseMatchupHandler
             return;
         }
 
-        // Fetch all other players in the queue
-        $opponents = $this->quickMatchService->fetchQmQueueEntry($this->history, $this->qmQueueEntry);
+        // Fetch all other players in the queue and exclude current player
+        $opponents = $this->quickMatchService->fetchQmQueueEntry($this->history, $this->qmQueueEntry)
+            ->filter(function($entry) {
+                return $entry->id !== $this->qmQueueEntry->id;
+            })->values();
         $count = $opponents->count() + 1;
         $timeInQueue = $this->qmQueueEntry->secondsinQueue();
         Log::debug("FindOpponent ** inQueue={$playerInQueue}, players in q: {$count}, for ladder={$ladder->abbreviation}, seconds in Queue: {$timeInQueue}");
@@ -35,10 +39,13 @@ class TeamMatchupHandler extends BaseMatchupHandler
         $matchableOpponents = $this->quickMatchService->getEntriesInSameTier($ladder, $this->qmQueueEntry, $opponents);
 
         // Find opponents that can be matched with current player.
-        $matchableOpponents = $this->quickMatchService->getEntriesInPointRange($this->qmQueueEntry, $matchableOpponents);
-      
+        $matchableOpponents = $this->getEntriesInPointRange2v2($this->qmQueueEntry, $matchableOpponents);
+
         $opponentCount = $matchableOpponents->count();
-        Log::debug("FindOpponent ** inQueue={$playerInQueue}, amount of matchable opponent after point filter: {$opponentCount} of {$count}");
+        $matchableNames = $matchableOpponents->map(function($entry) {
+            return $entry->qmPlayer?->player?->username ?? 'Unknown';
+        })->implode(', ');
+        Log::debug("FindOpponent ** inQueue={$playerInQueue}, amount of matchable opponent after point filter: {$opponentCount} of {$count}. Names: [{$matchableNames}]");
 
         // Count the number of players we need to start a match
         // Excluding current player
@@ -63,6 +70,21 @@ class TeamMatchupHandler extends BaseMatchupHandler
             . json_encode($teamAPlayers) . ' VS'
             . json_encode($teamBPlayers));
 
+        // Log player names and their team
+        $teamALog = $teamAPlayers->map(function($entry) {
+            return $entry->qmPlayer?->player?->username ?? 'Unknown';
+        })->implode(', ');
+        $teamBLog = $teamBPlayers->map(function($entry) {
+            return $entry->qmPlayer?->player?->username ?? 'Unknown';
+        })->implode(', ');
+        Log::debug("Team A (" . $teamAPlayers->count() . "): " . $teamALog);
+        Log::debug("Team B (" . $teamBPlayers->count() . "): " . $teamBLog);
+
+        // Ensure both teams have exactly two players
+        if ($teamAPlayers->count() !== 2 || $teamBPlayers->count() !== 2) {
+            Log::warning("Team size error: Team A has {$teamAPlayers->count()} players, Team B has {$teamBPlayers->count()} players. Expected 2 each.");
+        }
+
         $players = $teamAPlayers->merge($teamBPlayers);
 
         $commonQmMaps = $this->quickMatchService->getCommonMapsForPlayers($ladder, $players);
@@ -76,9 +98,15 @@ class TeamMatchupHandler extends BaseMatchupHandler
 
         // Add observers to the match if there is any
         $observers = $opponents->filter(fn(QmQueueEntry $qmQueueEntry) => $qmQueueEntry->qmPlayer?->isObserver());
-        if ($observers->count() < 0)
+        if ($observers->count() > 0)
         {
             $this->matchHasObservers = true;
+        }
+
+        // Throw exception if team sizes are not exactly two
+        if ($teamAPlayers->count() !== 2 || $teamBPlayers->count() !== 2) {
+            Log::warning("Team size error: Team A has {$teamAPlayers->count()} players, Team B has {$teamBPlayers->count()} players. Expected 2 each.");
+            throw new \RuntimeException("Team size error: Team A has {$teamAPlayers->count()} players, Team B has {$teamBPlayers->count()} players. Expected 2 each.");
         }
 
         // Start the match with all other players and other observers if there is any
@@ -107,5 +135,231 @@ class TeamMatchupHandler extends BaseMatchupHandler
             Game::GAME_TYPE_2VS2,
             $stats
         );
+    }
+
+    /**
+     * Attempt to find a valid 2v2 match for the given player.
+     * Ensures that:
+     * - The current player is always included in the match.
+     * - All players in the match are within each other's allowed point range.
+     * - The result is sorted by closeness in point range to improve match fairness.
+     *
+     * @param QmQueueEntry $currentQmQueueEntry
+     * @param Collection|QmQueueEntry[] $opponents
+     * @return Collection|QmQueueEntry[]  A collection containing the matched players (including current), or empty if no match found
+     */
+    public function getEntriesInPointRange2v2(QmQueueEntry $currentQmQueueEntry, Collection $opponents): Collection
+    {
+        $rules = $currentQmQueueEntry->ladderHistory->ladder->qmLadderRules;
+        $playerName = $currentQmQueueEntry->qmPlayer?->player?->username ?? 'Unknown';
+
+        Log::debug("2v2 Search start: queueEntry={$currentQmQueueEntry->id}, name={$playerName}, opponentsInQueue=" . count($opponents));
+
+        // Filter opponents to those that pass point range check with the current player
+        $potentialOpponents = $this->filterOpponentsInRange($currentQmQueueEntry, $opponents, $rules)
+            ->filter(function($opponent) use ($currentQmQueueEntry) {
+                return $opponent->id !== $currentQmQueueEntry->id;
+            })->values();
+
+        Log::debug($potentialOpponents->count() . " potential opponents for {$playerName}: " . $potentialOpponents->pluck('qmPlayer.player.username')->implode(', '));
+
+        if ($potentialOpponents->count() < 3)
+        {
+            Log::debug("Not enough valid opponents for {$playerName} to form a 2v2 match: " . $potentialOpponents->count() . "/3");
+            return collect();
+        }
+
+        // Sort opponents by a combination of point difference and time in queue bonus
+        $pointsPerSecond = $rules->points_per_second;
+        $sortedOpponents = $potentialOpponents->sortBy(function ($opponent) use ($currentQmQueueEntry, $pointsPerSecond) {
+            $pointDiff = abs($currentQmQueueEntry->points - $opponent->points);
+            $waitTimeBonus = (strtotime($opponent->updated_at) - strtotime($opponent->created_at)) * $pointsPerSecond;
+            return $pointDiff - $waitTimeBonus;
+        })->values();
+
+        // Try all possible 3-player combinations (since the current player makes 4)
+        foreach ($this->getCombinations($sortedOpponents, 3) as $threeOthers)
+        {
+            $matchPlayers = collect([$currentQmQueueEntry])->merge($threeOthers)->unique('id')->values();
+
+            if ($matchPlayers->count() !== 4) {
+                continue;
+            }
+
+            if ($this->allPlayersInRange($matchPlayers, $rules))
+            {
+                Log::debug(
+                    "✅ Found valid 2v2 match for {$playerName}: " .
+                        $matchPlayers->pluck('qmPlayer.player.username')->implode(', ')
+                );
+                return $matchPlayers;
+            }
+        }
+
+        Log::debug("❌ No valid 2v2 match found for {$playerName}");
+        return collect();
+    }
+
+    /**
+     * Generate combinations of a given size from a collection.
+     *
+     * @param Collection $items
+     * @param int $size
+     * @return Collection
+     */
+    private function getCombinations(Collection $items, int $size): Collection
+    {
+        $array = $items->all();
+        $allNames = collect($array)->map(function($entry) {
+            return $entry->qmPlayer?->player?->username ?? 'Unknown';
+        })->implode(', ');
+        Log::debug("getCombinations called with items: [{$allNames}] and size: {$size}");
+        $results = [];
+
+        $recurse = function ($arr, $size, $start = 0, $current = []) use (&$results, &$recurse)
+        {
+            if (count($current) === $size)
+            {
+                // Debug log the combination
+                $names = collect($current)->map(function($entry) {
+                    return $entry->qmPlayer?->player?->username ?? 'Unknown';
+                })->implode(', ');
+                Log::debug("getCombinations: [{$names}]");
+                $results[] = $current;
+                return;
+            }
+            for ($i = $start; $i < count($arr); $i++)
+            {
+                $current[] = $arr[$i];
+                $recurse($arr, $size, $i + 1, $current);
+                array_pop($current);
+            }
+        };
+
+        $recurse($array, $size);
+
+        return collect($results);
+    }
+
+    /**
+     * Filter opponents that are within point range of the current player.
+     *
+     * @param QmQueueEntry $current
+     * @param Collection|QmQueueEntry[] $opponents
+     * @param QmLadderRules $rules
+     * @return Collection|QmQueueEntry[]
+     */
+    private function filterOpponentsInRange(QmQueueEntry $current, Collection $opponents, QmLadderRules $rules): Collection
+    {
+        $pointsPerSecond = $rules->points_per_second;
+        $maxPointsDifference = $rules->max_points_difference;
+        $currentPointFilter = $current->qmPlayer->player->user->userSettings->disabledPointFilter;
+
+        $matchable = collect();
+
+        foreach ($opponents as $opponent)
+        {
+            if (!isset($opponent->qmPlayer) || $opponent->qmPlayer->isObserver())
+            {
+                continue;
+            }
+
+            $diff = abs($current->points - $opponent->points);
+            $waitTimeBonus = (strtotime($current->updated_at) - strtotime($current->created_at)) * $pointsPerSecond;
+
+            $inNormalRange = $waitTimeBonus + $maxPointsDifference > $diff;
+            $inDisabledFilterRange = $currentPointFilter
+                && $opponent->qmPlayer->player->user->userSettings->disabledPointFilter
+                && $diff < 1000
+                && $current->points > 400
+                && $opponent->points > 400;
+
+            if ($inNormalRange || $inDisabledFilterRange)
+            {
+                $matchable->push($opponent);
+            }
+        }
+
+        return $matchable;
+    }
+
+    /**
+     * Check if all players in the given collection are within point range of each other.
+     * Logs a pass/fail table for each comparison.
+     *
+     * @param Collection|QmQueueEntry[] $players
+     * @param QmLadderRules $rules
+     * @return bool
+     */
+    private function allPlayersInRange(Collection $players, QmLadderRules $rules): bool
+    {
+        $pointsPerSecond = $rules->points_per_second;
+        $maxPointsDifference = $rules->max_points_difference;
+        $comparisonResults = [];
+
+        foreach ($players as $i => $p1)
+        {
+            foreach ($players as $j => $p2)
+            {
+                if ($i >= $j) continue; // Skip self and duplicate checks
+
+                $diff = abs($p1->points - $p2->points);
+                $waitTimeBonusP1 = (strtotime($p1->updated_at) - strtotime($p1->created_at)) * $pointsPerSecond;
+                $waitTimeBonusP2 = (strtotime($p2->updated_at) - strtotime($p2->created_at)) * $pointsPerSecond;
+
+                $passesNormalRange = ($waitTimeBonusP1 + $maxPointsDifference >= $diff)
+                    || ($waitTimeBonusP2 + $maxPointsDifference >= $diff);
+
+                $passesDisabledFilter = $p1->qmPlayer->player->user->userSettings->disabledPointFilter
+                    && $p2->qmPlayer->player->user->userSettings->disabledPointFilter
+                    && $diff < 1000
+                    && $p1->points > 400
+                    && $p2->points > 400;
+
+                $pass = $passesNormalRange || $passesDisabledFilter;
+
+                $comparisonResults[] = [
+                    'p1' => $p1->qmPlayer?->player?->username ?? 'Unknown',
+                    'p2' => $p2->qmPlayer?->player?->username ?? 'Unknown',
+                    'points1' => $p1->points,
+                    'points2' => $p2->points,
+                    'diff' => $diff,
+                    'pass' => $pass
+                ];
+
+                if (!$pass)
+                {
+                    $this->logComparisonTable($comparisonResults);
+                    return false;
+                }
+            }
+        }
+
+        $this->logComparisonTable($comparisonResults);
+        return true;
+    }
+
+    /**
+     * Logs the comparison results for all player pairs.
+     *
+     * @param array $comparisonResults
+     */
+    private function logComparisonTable(array $comparisonResults): void
+    {
+        Log::debug("=== 2v2 Player Comparison Table ===");
+        foreach ($comparisonResults as $result)
+        {
+            $status = $result['pass'] ? '✅ PASS' : '❌ FAIL';
+            Log::debug(sprintf(
+                "%-15s (%4d pts) ↔ %-15s (%4d pts) | Diff: %4d | %s",
+                $result['p1'],
+                $result['points1'],
+                $result['p2'],
+                $result['points2'],
+                $result['diff'],
+                $status
+            ));
+        }
+        Log::debug("===================================");
     }
 }
