@@ -10,6 +10,7 @@ use Exception;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Validation\ValidationException;
+use Illuminate\Support\Facades\Log;
 
 class ApiPlayerController extends Controller
 {
@@ -73,6 +74,7 @@ class ApiPlayerController extends Controller
         }
         catch (Exception $ex)
         {
+            Log::error("Exception in ApiPlayerController@getUsernames", ['exception' => $ex]);
             return response()->json(["message" => "Something went wrong"], 500);
         }
     }
@@ -100,7 +102,7 @@ class ApiPlayerController extends Controller
 
             # Check ladder exists
             $ladder = \App\Models\Ladder::where("abbreviation", '=', $request->ladderAbbrev)->first();
-            $ladderHistory = $ladder->currentHistory;
+            $ladderHistory = $ladder->currentHistory();
             if ($ladder === null || $ladderHistory === null)
             {
                 return response()->json(["message" => "Ladder does not exist"], 400);
@@ -126,10 +128,12 @@ class ApiPlayerController extends Controller
         }
         catch (ValidationException $ex)
         {
+            Log::warning("ValidationException in ApiPlayerController@createPlayer", ['errors' => $ex->errors()]);
             return response()->json(["message" => $ex->getMessage()], 400);
         }
         catch (Exception $ex)
         {
+            Log::error("Exception in ApiPlayerController@createPlayer", ['exception' => $ex]);
             return response()->json(["message" => "Something went wrong"], 500);
         }
     }
@@ -221,10 +225,12 @@ class ApiPlayerController extends Controller
         }
         catch (ValidationException $ex)
         {
+            Log::warning("ValidationException in ApiPlayerController@togglePlayerStatus", ['errors' => $ex->errors()]);
             return response()->json(["message" => $ex->getMessage()], 400);
         }
         catch (Exception $ex)
         {
+            Log::error("Exception in ApiPlayerController@togglePlayerStatus", ['exception' => $ex]);
             return response()->json(["message" => "Something went wrong"], 500);
         }
     }
