@@ -16,6 +16,7 @@ use App\Models\Player;
 use App\Models\User;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
+use \App\Models\PlayerActiveHandle;
 
 class LadderController extends Controller
 {
@@ -225,13 +226,13 @@ class LadderController extends Controller
         }
         else
         {
-            $allGameReports = $game->report()->get();
+            $allGameReports = $game->report;
             $userIsMod = false;
         }
 
         if ($reportId !== null)
         {
-            $gameReport = $game->allReports()->where('game_reports.id', '=', $reportId)->first();
+            $gameReport = $game->allReports->where('id', $reportId)->first();
         }
         else
         {
@@ -256,6 +257,28 @@ class LadderController extends Controller
 
         $playerGameReports = $gameReport->playerGameReports ?? [];
         $groupedPlayerGameReports = [];
+
+        $showBothPositiveFix = false;
+        $showBothZeroFix = false;
+        $fixedPointsPreview = [];
+
+        if (count($playerGameReports) === 2)
+        {
+            $p1 = $playerGameReports[0];
+            $p2 = $playerGameReports[1];
+
+            $hasOneWinner = $p1->won != $p2->won;
+            $bothPositive = $p1->points > 0 && $p2->points > 0;
+            $bothZero = $p1->points == 0 && $p2->points == 0;
+
+            $showBothPositiveFix = $hasOneWinner && $bothPositive;
+            $showBothZeroFix = $hasOneWinner && $bothZero;
+
+            if ($showBothZeroFix && $userIsMod)
+            {
+                $fixedPointsPreview = app(\App\Http\Controllers\AdminController::class)->awardedPointsPreview($gameReport, $history);
+            }
+        }
 
         if ($history->ladder->ladder_type == Ladder::TWO_VS_TWO)
         {
@@ -349,6 +372,9 @@ class LadderController extends Controller
                     "qmMatchPlayers" => $qmMatchPlayers,
                     "tunnels" => $tunnels,
                     "date" => $date,
+                    "showBothPositiveFix" => $showBothPositiveFix,
+                    "showBothZeroFix" => $showBothZeroFix,
+                    "fixedPointsPreview" => $fixedPointsPreview,
                 ]
             );
         }
@@ -375,6 +401,9 @@ class LadderController extends Controller
                     "qmConnectionStats" => $qmConnectionStats,
                     "qmMatchPlayers" => $qmMatchPlayers,
                     "date" => $date,
+                    "showBothPositiveFix" => $showBothPositiveFix,
+                    "showBothZeroFix" => $showBothZeroFix,
+                    "fixedPointsPreview" => $fixedPointsPreview,
                 ]
             );
         }
@@ -383,7 +412,6 @@ class LadderController extends Controller
     public function getLadderPlayer(Request $request, $date = null, $cncnetGame = null, $username = null)
     {
         $history = $this->ladderService->getActiveLadderByDate($date, $cncnetGame);
-        $currentHistory = $history->ladder->currentHistory();
 
         if ($history == null)
         {
@@ -450,17 +478,32 @@ class LadderController extends Controller
         $recentAchievements = $this->achievementService->getRecentlyUnlockedAchievements($history, $user, 3);
         $achievementProgressCounts = $this->achievementService->getProgressCountsByUser($history, $user);
 
-        $isAnonymous = $player->user->userSettings->is_anonymous;
+        $isAnonymous = $player->user->userSettings->getIsAnonymousForLadderHistory($history);
 
-        $ladderNicks = [];
-        if (!$isAnonymous && $history->id != $currentHistory->id) // only hide if anonymous and is the current month
-        {
-            $ladderNicks = $user->usernames
-                ->where('id', '!=', $player->id)
-                ->where('ladder_id', $history->ladder->id)
-                ->pluck('username')
-                ->toArray();
+        // Get the current month's date range
+        $now = Carbon::now();
+        $dateStart = $now->copy()->startOfMonth()->toDateTimeString();
+        $dateEnd = $now->copy()->endOfMonth()->toDateTimeString();
+
+        // Get the player's active handle for the current month
+        $activeHandle = PlayerActiveHandle::getPlayerActiveHandle(
+            $player->id,
+            $history->ladder->id,
+            $dateStart,
+            $dateEnd
+        );
+
+        // Start with all usernames for this ladder, excluding the current player
+        $ladderNicks = $user->usernames
+            ->where('id', '!=', $player->id)
+            ->where('ladder_id', $history->ladder->id);
+
+        // If anonymous and there's an active handle, also exclude it
+        if ($isAnonymous && $activeHandle) {
+            $ladderNicks = $ladderNicks->where('id', '!=', $activeHandle->player_id);
         }
+
+        $ladderNicks = $ladderNicks->pluck('username')->toArray();
 
         return view(
             "ladders.player-detail",
