@@ -24,6 +24,7 @@ class ChartService
                 ->get();
 
             $labels = [];
+            $results = [];
             foreach ($games as $hour => $game)
             {
                 $hourFormatted = Carbon::create(null, null, null, $hour);
@@ -42,7 +43,7 @@ class ChartService
 
     public function getPlayerGamesPlayedByMonth($player, $history)
     {
-        return Cache::remember("getPlayerGamesPlayedByMonth/$history->short/$player->id", 5 * 60, function () use ($player, $history)
+        return Cache::remember("getPlayerGamesPlayedByMonth/$history->short/$player->id", 10 * 60, function () use ($player, $history)
         {
             $now = $history->starts;
             $from = $now->copy()->startOfMonth()->toDateTimeString();
