@@ -36,23 +36,27 @@ class QuickMatchSpawnService
 
         srand($qmMatch->seed); // Seed the RNG for possibly random boolean options
 
+        $isObserver = $qmPlayer->is_observer == 1;
+        $notAllowedToChat = !$qmPlayer->player->user->getIsAllowedToChat();
+
         $spawnStruct["spawn"]["Settings"] = array_filter(
             [
                 "UIGameMode" =>     $qmMap->game_mode,
                 "UIMapName" =>      $qmMap->description,
                 "MapHash" =>        $map->hash,
-                "Scenario" =>       ($map->filename && $ladder->game == 'd2k') ? $map->filename : "spawnmap.ini",
+                "Scenario" => ($map->filename && $ladder->game == 'd2k') ? $map->filename : "spawnmap.ini",
                 "Seed" =>           $qmMatch->seed,
                 "GameID" =>         $qmMatch->seed,
                 "WOLGameID" =>      $qmMatch->seed,
-                "Host" =>           ($qmPlayer->color == 0 && $ladder->abbreviation == "d2k") ? "Yes" : "No", // if Dune and player color is 0
+                "Host" => ($qmPlayer->color == 0 && $ladder->abbreviation == "d2k") ? "Yes" : "No", // if Dune and player color is 0
                 "Name" =>           $qmPlayer->player()->first()->username,
                 "Port" =>           $qmPlayer->port,
                 "Side" =>           $qmPlayer->actual_side,
                 "Color" =>          $qmPlayer->color,
                 "MyIndex" =>        $qmPlayer->color,
                 "IsSpectator" =>    "False",
-                "DisableChat" =>    $qmPlayer->player->user->getIsAllowedToChat() ? "False" : "True",
+                "DisableChat" => ($isObserver || $notAllowedToChat) ? "True" : "False",
+                "AllowChat" => (!$isObserver && !$notAllowedToChat) ? "True" : "False",
                 // Filter null values
             ],
             function ($var)
@@ -121,7 +125,10 @@ class QuickMatchSpawnService
      */
     public static function appendOthersAndTeamAlliancesToSpawnIni($spawnStruct, $qmPlayer, $otherQmMatchPlayers)
     {
-        Log::debug("QuickMatchSpawnService ** appendOthersAndTeamAlliancesToSpawnIni: Is Observer: " . $qmPlayer->isObserver());
+        Log::debug("QuickMatchSpawnService ** appendOthersAndTeamAlliancesToSpawnIni", [
+            'username' => $qmPlayer->player?->username ?? 'N/A',
+            'isObserver' => $qmPlayer->isObserver() ? true : false
+        ]);
 
         $otherIndex = 1;
         $multiIndex = $qmPlayer->color + 1;
@@ -347,7 +354,7 @@ class QuickMatchSpawnService
     {
 
         // group all players by team
-        $playersByTeam = $otherQmMatchPlayers->concat([$qmPlayer])->groupBy(fn ($p) => $p->team);
+        $playersByTeam = $otherQmMatchPlayers->concat([$qmPlayer])->groupBy(fn($p) => $p->team);
         foreach ($playersByTeam as $team => $players)
         {
             for ($i = 0; $i < $players->count(); $i++)
@@ -374,7 +381,7 @@ class QuickMatchSpawnService
         $RA1_HOUSE_MULTI_INDEX_OFFSET = 12;
 
         // Group all players by team
-        $playersByTeam = $otherQmMatchPlayers->concat([$qmPlayer])->groupBy(fn ($p) => $p->team);
+        $playersByTeam = $otherQmMatchPlayers->concat([$qmPlayer])->groupBy(fn($p) => $p->team);
         foreach ($playersByTeam as $team => $players)
         {
             for ($i = 0; $i < $players->count(); $i++)
@@ -414,27 +421,34 @@ class QuickMatchSpawnService
      */
     public static function appendObservers($spawnStruct, $qmPlayer, $otherQmMatchPlayers)
     {
-        # Checks if current player is observer
+        // Sort players by color
+        $allPlayers = $otherQmMatchPlayers->concat([$qmPlayer])->sortBy('color')->values();
+
+        $rankByPlayerId = array();
+        $index = 0;
+        foreach ($allPlayers as $player)
+        {
+            $rankByPlayerId[$player->id] = $index + 1;
+            $index = $index + 1;
+        }
+
         if ($qmPlayer->isObserver())
         {
-            $playerIndex = 1;
             $spawnStruct["spawn"]["Settings"]["IsSpectator"] = "True";
-            $spawnStruct["isspectator"]["Multi$playerIndex"] = "True";
         }
 
-        # Make sure we mark other players too
-        foreach ($otherQmMatchPlayers as $playerIndex => $opn)
+        // Multi<Rank> = True for every observer
+        foreach ($allPlayers as $player)
         {
-            if ($opn->isObserver())
+            if ($player->isObserver())
             {
-                # Because it references "Other", which is 1-8
-                $playerIndex = $playerIndex + 1;
-                $spawnStruct["isspectator"]["Multi$playerIndex"] = "True";
+                $rank = $rankByPlayerId[$player->id];
+                $spawnStruct["isspectator"]["Multi" . $rank] = "True";
             }
         }
+
         return $spawnStruct;
     }
-
 
     /**
      * Prepend quick-coop ini file to allow 2 real players vs 2 ai
