@@ -52,6 +52,11 @@ class QmMatch extends Model
 
     public function findQmPlayerByPlayerId(int $playerId)
     {
-        return $this->players->where('player_id', '=', $playerId)->first();
+        return $this->players->where('player_id', '=', $playerId)->first(); // returns a qm_match_player
+    }
+
+    public function observers()
+    {
+        return $this->hasMany(QmMatchPlayer::class)->where('is_observer', 1);
     }
 }
