@@ -19,15 +19,17 @@ class StatsCache
         $from = $now->copy()->startOfDay()->toDateTimeString();
         $to = $now->copy()->endOfDay()->toDateTimeString();
 
-        $playersToday =  PlayerGameReport::join("game_reports", "game_reports.id", "=", "player_game_reports.game_report_id")
-            ->join("games", "games.id", "=", "game_reports.game_id")
-            ->whereBetween("player_game_reports.created_at", [$from, $to])
-            ->where("games.ladder_history_id", $history->id)
-            ->where("game_reports.valid", true)
-            ->groupBy("player_game_reports.player_id")
-            ->pluck("player_game_reports.player_id");
+        // TODO fix slow performance
 
-        Cache::put("playersToday.$history->id", $playersToday, 1800 * 60);
+        // $playersToday =  PlayerGameReport::join("game_reports", "game_reports.id", "=", "player_game_reports.game_report_id")
+        //     ->join("games", "games.id", "=", "game_reports.game_id")
+        //     ->whereBetween("player_game_reports.created_at", [$from, $to])
+        //     ->where("games.ladder_history_id", $history->id)
+        //     ->where("game_reports.valid", true)
+        //     ->groupBy("player_game_reports.player_id")
+        //     ->pluck("player_game_reports.player_id");
+
+        // Cache::put("playersToday.$history->id", $playersToday, 1800 * 60);
     }
 
 
