@@ -28,7 +28,7 @@ return [
 		],
 
 		'api' => [
-			'driver' => 'token',
+			'driver' => 'jwt',
 			'provider' => 'users',
 		],
 	],
