@@ -16,7 +16,7 @@ return [
 	|
 	*/
 
-    'default' => env('QUEUE_DRIVER', 'database'),
+    'default' => env('QUEUE_DRIVER', 'redis'),
 
 	/*
 	|--------------------------------------------------------------------------
@@ -43,9 +43,12 @@ return [
 		],
 
 		'redis' => [
-			'driver' => 'redis',
-			'queue'  => 'default',
-			'retry_after' => 60,
+            'driver' => 'redis',
+            'connection' => env('REDIS_QUEUE_CONNECTION', 'default'),
+            'queue' => env('REDIS_QUEUE', 'default'),
+            'retry_after' => (int) env('REDIS_QUEUE_RETRY_AFTER', 1800),
+            'block_for' => null,
+            'after_commit' => false,
 		],
 
 	],
