@@ -2,9 +2,10 @@
 <div class="bans">
     <h5>Bans:</h5>
 
-    <?php foreach($user->bans as $ban):?>
+    <?php foreach($user->collectBans() as $ban):?>
         <div style="margin-bottom: 15px;">
         <ul class="list-styled">
+            <li><strong>User ID:</strong> #{{ $ban->user_id }}</li>
             <li><strong>Internal note:</strong> {{ $ban->internal_note }}</li>
             <li><strong>Public reason:</strong> {{ $ban->plubic_reason }}</li>
             <li><strong>Expires:</strong> {{ $ban->expires->toDateString() }}</li>
