@@ -109,18 +109,21 @@
                 <div>
                     Elo: @@elo@@
                 </div>
-                <div>
+                <div class="js-elo-rank">
                     Rank: #@@elo_rank@@
                 </div>
                 <div>
                     Games: @@elo_games@@
                 </div>
-                <div>
-                    Deviation: @@elo_deviation@@
-                </div>
             </div>
         </div>
         <div class="player-info-row">
+            <div class="player-info-col js-alias">
+                <h5 class="title">Alias</h5>
+                <div class="last-online">
+                    @@alias@@
+                </div>
+            </div>
             <div class="player-info-col js-joined">
                 <h5 class="title">User Joined</h5>
                 <div class="last-online">
@@ -130,93 +133,102 @@
         </div>
     </div>
 
-    @if ($ladderHasEnded == false)
-        <script>
-            (function() {
-                let trigger = document.getElementById("js_profile_{{ $username }}");
-                let template = document.getElementById("js_template_{{ $username }}");
-                let player = "{{ $username }}";
-                let ladder = "{{ $history->ladder->abbreviation }}";
+    <script>
+        (function() {
+            let trigger = document.getElementById("js_profile_{{ $username }}");
+            let template = document.getElementById("js_template_{{ $username }}");
+            let player = "{{ $username }}";
+            let ladder = "{{ $history->ladder->abbreviation }}";
 
-                const tippyInstance = tippy(trigger, {
-                    allowHTML: true,
-                    theme: "player-card",
-                    placement: "auto-start",
-                    touch: false,
-                    maxWidth: 500,
-                    interactive: true,
-                    interactiveBorder: 10,
-                    content: getLoadingContent(), // Initial loading state
-                    onShow: () => fetchPlayer(),
-                });
+            const tippyInstance = tippy(trigger, {
+                allowHTML: true,
+                theme: "player-card",
+                placement: "auto-start",
+                touch: false,
+                maxWidth: 500,
+                interactive: true,
+                interactiveBorder: 10,
+                content: getLoadingContent(), // Initial loading state
+                onShow: () => fetchPlayer(),
+            });
 
-                // Create an object to hold the cached data
-                const cache = {};
+            // Create an object to hold the cached data
+            const cache = {};
 
-                function getLoadingContent() {
-                    return `<div class="loading-message">Loading...</div>`;
+            function getLoadingContent() {
+                return `<div class="loading-message">Loading...</div>`;
+            }
+
+            function fetchPlayer() {
+                // Check if data is already in cache
+                if (cache[player]) {
+                    // If cached data is available, update the tooltip content with it
+                    updateTippyContent(cache[player]);
+                } else {
+                    // If data is not in cache, make an API request to fetch the data
+                    fetchDataAndDisplay(player, ladder);
+                }
+            }
+
+            async function fetchDataAndDisplay(playerId, ladderId) {
+                try {
+                    const baseUrl = document.querySelector('meta[name="api-base-url"]')?.getAttribute('content') ?? '';
+                    let date = "{{ $history->starts->format('m-Y') }}";
+                    const apiUrl = `${baseUrl}/api/v1/ladder/${ladderId}/player/${playerId}?date=${date}`;
+                    const response = await fetch(apiUrl);
+                    if (!response.ok) {
+                        throw new Error('Network response was not ok');
+                    }
+                    const data = await response.json();
+                    console.log("Tooltip-Daten:", data);
+                    // Cache the fetched data
+                    cache[player] = data;
+
+                    // Update the Tippy.js tooltip content with the actual data
+                    updateTippyContent(data);
+                } catch (error) {
+                    // Display an error message if needed
+                    tippyInstance.setContent(`<div class="error-message">Error loading data</div>`);
+                }
+            }
+
+            function updateTippyContent(data) {
+                // Create the badge divs dynamically
+                let badgesHTML = '';
+                for (let i = 0; i < data.last_five_games?.length; i++) {
+                    const won = data.last_five_games[i].won;
+                    badgesHTML += `<div class="badge ${won ? 'badge-won' : 'badge-lost'}">${won ? "W" : "L"}</div>`;
                 }
 
-                function fetchPlayer() {
-                    // Check if data is already in cache
-                    if (cache[player]) {
-                        // If cached data is available, update the tooltip content with it
-                        updateTippyContent(cache[player]);
-                    } else {
-                        // If data is not in cache, make an API request to fetch the data
-                        fetchDataAndDisplay(player, ladder);
-                    }
+                if (!data.alias) {
+                    template.querySelector(".js-alias").style.display = "none";
                 }
 
-                async function fetchDataAndDisplay(playerId, ladderId) {
-                    try {
-                        const response = await fetch(`/api/v1/ladder/${ladderId}/player/${playerId}`);
-                        if (!response.ok) {
-                            throw new Error('Network response was not ok');
-                        }
-                        const data = await response.json();
-
-                        // Cache the fetched data
-                        cache[player] = data;
-
-                        // Update the Tippy.js tooltip content with the actual data
-                        updateTippyContent(data);
-                    } catch (error) {
-                        // Display an error message if needed
-                        tippyInstance.setContent(`<div class="error-message">Error loading data</div>`);
-                    }
+                if (data.elo == null) {
+                    template.querySelector(".js-elo").style.display = "none";
+                    template.querySelector(".js-elo-rank").style.display = "none";
+                }
+                else if (data.elo.elo_rank == 0) {
+                    template.querySelector(".js-elo-rank").style.display = "none";
                 }
 
-                function updateTippyContent(data) {
-                    // Create the badge divs dynamically
-                    let badgesHTML = '';
-                    for (let i = 0; i < data.last_five_games?.length; i++) {
-                        const won = data.last_five_games[i].won;
-                        badgesHTML += `<div class="badge ${won ? 'badge-won' : 'badge-lost'}">${won ? "W" : "L"}</div>`;
-                    }
-
-                    if (data.elo == null) {
-                        template.querySelector(".js-elo").style.display = "none";
-                    }
-
-                    if (data.user_since == null) {
-                        template.querySelector(".js-joined").style.display = "none";
-                    }
-
-                    // Replace the loading message with the actual badge divs and last_active value
-                    let content = template.innerHTML
-                        .replace("@@badges@@", badgesHTML)
-                        .replace("@@last_active@@", data.last_active ?? "")
-                        .replace("@@elo@@", data.elo?.elo ?? "")
-                        .replace("@@elo_rank@@", data.elo?.rank ?? "")
-                        .replace("@@elo_games@@", data.elo?.game_count ?? "")
-                        .replace("@@elo_deviation@@", data.elo?.deviation ?? "")
-                        .replace("@@user_since@@", data.user_since ?? "");
-
-                    // Update Tippy.js content
-                    tippyInstance.setContent(content);
+                if (data.user_since == null) {
+                    template.querySelector(".js-joined").style.display = "none";
                 }
-            })();
-        </script>
-    @endif
+
+                // Replace the loading message with the actual badge divs and last_active value
+                let content = template.innerHTML
+                    .replace("@@badges@@", badgesHTML)
+                    .replace("@@last_active@@", data.last_active ?? "")
+                    .replace("@@elo@@", data.elo ? (data.elo.rating + (data.elo.active ? "" : "?")) : "")
+                    .replace("@@elo_rank@@", data.elo?.elo_rank ?? "")
+                    .replace("@@elo_games@@", data.elo?.rated_games ?? "")
+                    .replace("@@user_since@@", data.user_since ?? "")
+                    .replace("@@alias@@", data.alias ?? "");
+
+                // Update Tippy.js content
+                tippyInstance.setContent(content);
+            }
+        })();
+    </script>
 </div>
