@@ -10,6 +10,7 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <meta name="csrf-token" content="{{ csrf_token() }}">
+    <meta name="api-base-url" content="{{ url('/') }}">
 
     <title>@yield('title') - CnCNet</title>
 
@@ -54,7 +55,7 @@
 <body class="@yield('body-class'){{ $bodyClass }} @hasSection('page-body-class')@yield('page-body-class')@endif">
 
     @if(isset($history))
-        @include('components.countdown', ['target' => $history->ends->toISO8601String()])
+        @include('components.countdown', ['target' => $history->ends?->toISO8601String()])
     @endif
 
     <a href="#" class="live-event js-live-event hidden" target="_blank" title="WATCH LIVE - RED ALERT 2 &amp; YURI'S REVENGE WORLD CHAMPIONSHIPS">
